refactor: 모달 관련 패키지 정리 (#52)
