+++ app/src/main/java/com/hyeeyoung/wishboard/designsystem/component/dialog/screen/ModalActivity.kt
@@ -1,4 +1,4 @@
-package com.hyeeyoung.wishboard.presentation.dialog
+package com.hyeeyoung.wishboard.designsystem.component.dialog.screen
 
 import android.content.Intent
 import android.os.Bundle
@@ -8,6 +8,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.SideEffect
 import androidx.compose.ui.graphics.Color
 import com.google.accompanist.systemuicontroller.rememberSystemUiController
+import com.hyeeyoung.wishboard.designsystem.component.dialog.model.ModalData
 import com.hyeeyoung.wishboard.presentation.folder.FolderListModalContent
 import com.hyeeyoung.wishboard.presentation.folder.FolderUploadModalContent
 import com.hyeeyoung.wishboard.presentation.noti.NotiModalContent
