+++ app/src/main/java/com/hyeeyoung/wishboard/designsystem/component/dialog/screen/WishBoardDialog.kt
@@ -1,4 +1,4 @@
-package com.hyeeyoung.wishboard.designsystem.component.dialog
+package com.hyeeyoung.wishboard.designsystem.component.dialog.screen
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -26,9 +26,9 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
+import com.hyeeyoung.wishboard.designsystem.component.dialog.model.DialogData
 import com.hyeeyoung.wishboard.designsystem.component.divider.WishBoardDivider
 import com.hyeeyoung.wishboard.designsystem.style.WishBoardTheme
-import com.hyeeyoung.wishboard.presentation.dialog.DialogData
 
 @Composable
 fun WishBoardDialog(
